Refactor RNG and game initialization; update seed handling to use uint64_t and enhance random number generation

+++ include/game_logic/game.h
@@ -15,7 +15,7 @@ class Game{
 	RNG rng;
 	EventLog event_log;
 
-	Game(std::deque<Player> &p, uint32_t s, int l);
+	Game(std::deque<Player> &p, uint64_t s, int l);
 	
 	void displayGameState(int floor, int turn);
